pyproject.toml
[tool.poetry]
name = "auto-manager"
version = "0.1.0"
description = ""
authors = ["geer.max0404 <geer.max0404@gmail.com>"]
readme = "README.md"
packages = [{include = "auto_manager"}]

[tool.poetry.dependencies]
python = "^3.12"
bestconfig = "^1.3.6"
scrapy = "^2.11.0"
mysql-connector-python = "^8.1.0"
pytz = "^2023.3.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
